Service destiny queries updated for sqlite conversion

+++ lib-levelcrush-rs/src/types/destiny.rs
@@ -4,7 +4,7 @@
 pub type MembershipId = i64;
 
 /// i32 meant to represent a membership type
-pub type MembershipType = i32;
+pub type MembershipType = i64;
 
 /// i64 menat to represent an instance
 pub type InstanceId = i64;
@@ -16,4 +16,4 @@ pub type CharacterId = i64;
 pub type GroupId = i64;
 
 /// u32 meant to represent a hash that comes from the destiny 2 manifest
-pub type ManifestHash = u32;
+pub type ManifestHash = i64;
